requirements.txt
streamlit>=1.32.0
yfinance>=0.2.37
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
plotly>=5.19.0
anthropic>=0.25.0
groq>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
fpdf2>=2.7.9
